Safari-611.1.10/LayoutTests/fast/dom/replace-child-with-mutation-event-removal-and-circular-template-insertion-expected.txt

This tests re-parenting a child of the document fragment during an insertion so as to create a circular node tree.
WebKit should detect this case and throw HierarchyRequestError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS container.replaceChild(fragment, refChild) threw exception HierarchyRequestError: The operation would yield an incorrect node tree..
PASS container.parentNode is child.content
PASS child.parentNode is null
PASS successfullyParsed is true

TEST COMPLETE
